requirements.txt
telethon
python-dotenv
PySocks
cryptg
aiohttp
psycopg2-binary
telethon-session-sqlalchemy
sqlalchemy
emoji
pillow
